+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,13 @@
 ## 1.0.0
 
 - Initial release
+
+## 1.0.1
+
+- Changed README to say it spawns inside. There was a small mistake
+- Fixed an error in the console where it came to finding nodes
+- Changed Rarity to 25
+- Broke the infinite loop iteration for teleport
+- Made multiplayer more stable
+- Improved Collider check for a better experience
+  - We now have a big vision check which make sure you don't look at 173 in any way. Not just the middle of his position